fix(ci): python-coverage-shards — split 1787-line overlay-retry test over the 1500-line maintainability limit
The maintainability guardrail test
test_first_party_code_files_stay_under_line_limit flagged
tests/unit/control/test_cleanup_auth_overlay_retry.py at 1787 lines,
over the 1500-line cap. Split it into a test_*_parts/ package following
the repo convention: part_001 keeps the in-memory control-flow doubles
and the shared _candidate/_RecordingLog/_REPO/_BASE helpers; part_002
holds the Postgres-backed candidate-query/record-helper coverage and
reuses those helpers via import. No test behavior changed (41 tests).
